updated tatoeba preprocessing functions

+++ tatoeba/preprocess.py
@@ -39,46 +39,47 @@ def get_tatoeba(url: str, force: bool = False):
     for split in split_list:
         out_file = DATA_FOLDER / f"deu-kor.{split}.csv"
 
-        id_file = open(
-            DATA_FOLDER / f"release/v2021-08-07/deu-kor/{split}.id",
-            "r",
-            encoding="utf-8",
-        )
+        if not out_file.exists() or force:
+            id_file = open(
+                DATA_FOLDER / f"release/v2021-08-07/deu-kor/{split}.id",
+                "r",
+                encoding="utf-8",
+            )
 
-        num_lines = _get_file_lines(id_file)
+            num_lines = _get_file_lines(id_file)
 
-        src_file = open(
-            DATA_FOLDER / f"release/v2021-08-07/deu-kor/{split}.src",
-            "r",
-            encoding="utf-8",
-        )
-        trg_file = open(
-            DATA_FOLDER / f"release/v2021-08-07/deu-kor/{split}.trg",
-            "r",
-            encoding="utf-8",
-        )
+            src_file = open(
+                DATA_FOLDER / f"release/v2021-08-07/deu-kor/{split}.src",
+                "r",
+                encoding="utf-8",
+            )
+            trg_file = open(
+                DATA_FOLDER / f"release/v2021-08-07/deu-kor/{split}.trg",
+                "r",
+                encoding="utf-8",
+            )
 
-        desc = f"generating {split} split as csv"
+            desc = f"generating {split} split as csv"
 
-        with tqdm(total=num_lines, desc=desc) as pbar:
-            with open(out_file, "w", encoding="utf-8", newline="") as out:
-                out_writer = csv.writer(out)
-                out_writer.writerow(["id", "source", "target"])
+            with tqdm(total=num_lines, desc=desc) as pbar:
+                with open(out_file, "w", encoding="utf-8", newline="") as out:
+                    out_writer = csv.writer(out)
+                    out_writer.writerow(["id", "source", "target"])
 
-                while True:
-                    idx = id_file.readline().strip()
-                    src = src_file.readline().strip()
-                    trg = trg_file.readline().strip()
+                    while True:
+                        idx = id_file.readline().strip()
+                        src = src_file.readline().strip()
+                        trg = trg_file.readline().strip()
 
-                    if not idx:
-                        break
+                        if not idx:
+                            break
 
-                    out_writer.writerow([idx, src, trg])
-                    pbar.update()
+                        out_writer.writerow([idx, src, trg])
+                        pbar.update()
 
-        id_file.close()
-        src_file.close()
-        trg_file.close()
+            id_file.close()
+            src_file.close()
+            trg_file.close()
 
 
 def get_dataset(force_renew: bool = False):
@@ -116,7 +117,7 @@ def get_subtitle_dataset(force_renew: bool = False):
 
         subsets = ("OpenSubtitles-v2018", "TED2020-v1")
 
-        dataset = get_dataset()["train"]
+        dataset = get_dataset(force_renew)["train"]
 
         subtitle_set = dataset.filter(
             lambda ex: ex["id"].startswith(subsets),
